+++ libft/libft.h
@@ -54,7 +54,7 @@ char		*ft_strrchr(const char *s, int c);
 size_t		ft_strlcpy(char *dest, const char *src, size_t size);
 size_t		ft_strlcat(char *dest, const char *src, size_t size);
 char		*ft_strnstr(const char *s1, const char *s2, size_t n);
-int			ft_atoi(const char *str);
+long long	ft_atoi(const char *str);
 void		*ft_calloc(size_t count, size_t size);
 
 /*
